Refactor service placement and add claim for id
