+++ TTE.API/Configurations/ApplicationServiceRegistration.cs
@@ -1,5 +1,6 @@
 ﻿using TTE.Application.Interfaces;
 using TTE.Application.Services;
+using TTE.Commons.Services;
 using TTE.Infrastructure.Repositories;
 
 namespace TTE.API.Configurations
@@ -9,8 +10,7 @@ namespace TTE.API.Configurations
         public static IServiceCollection AddApplicationServices(this IServiceCollection services)
         {
             services.AddScoped<IProductService, ProductService>();
-            services.AddScoped<IJwtService, JwtService>();
-            services.AddScoped<IPasswordHasherService, PasswordHasherService>();
+            services.AddScoped<ISecurityService, SecurityService>();
             services.AddScoped<IAuthService, AuthService>();
 
             services.AddScoped(typeof(IGenericRepository<>), typeof(GenericRepository<>));
