--- TTE.Application/Interfaces/IJwtService.cs
@@ -1,7 +1,0 @@
-﻿namespace TTE.Application.Interfaces
-{
-    public interface IJwtService
-    {
-        string GenerateToken(string username, string role);
-    }
-}
