--- TTE.Application/Interfaces/IPasswordHasherService.cs
@@ -1,8 +1,0 @@
-﻿namespace TTE.Application.Interfaces
-{
-    public interface IPasswordHasherService
-    {
-        string HashPassword(string password);
-        bool VerifyPassword(string password, string hashedPassword);
-    }
-}
