+++ TTE.Application/Services/AuthService.cs
@@ -1,5 +1,6 @@
 ﻿using TTE.Application.DTOs;
 using TTE.Application.Interfaces;
+using TTE.Commons.Services;
 using TTE.Infrastructure.Models;
 using TTE.Infrastructure.Repositories;
 
@@ -8,24 +9,22 @@ namespace TTE.Application.Services
     public class AuthService: IAuthService
     {
         private readonly IGenericRepository<User> _userRepository;
-        private readonly IPasswordHasherService _passwordHasher;
-        private readonly IJwtService _jwtService;
+        private readonly ISecurityService _securityService;
 
-        public AuthService(IGenericRepository<User> userRepository, IPasswordHasherService passwordHasher, IJwtService jwtService)
+        public AuthService(IGenericRepository<User> userRepository, ISecurityService securityService)
         {
             _userRepository = userRepository;
-            _passwordHasher = passwordHasher;
-            _jwtService = jwtService;
+            _securityService = securityService;
         }
 
         public async Task<LoginResponseDto?> LoginAsync(LoginRequestDto loginRequest)
         {
             var user = await _userRepository.GetByCondition(u => u.Email == loginRequest.Email, "Role");
 
-            if (user == null || !_passwordHasher.VerifyPassword(loginRequest.Password, user.Password))
+            if (user == null || !_securityService.VerifyPassword(loginRequest.Password, user.Password))
                 return null;
 
-            var token = _jwtService.GenerateToken(user.UserName, user.Role.Name);
+            var token = _securityService.GenerateToken(user.UserName, user.Role.Name, user.Id);
 
             return new LoginResponseDto
             {
