--- TTE.Application/Services/PasswordHasherService.cs
@@ -1,18 +1,0 @@
-﻿using TTE.Application.Interfaces;
-using BCrypt.Net;
-
-namespace TTE.Application.Services
-{
-    public class PasswordHasherService: IPasswordHasherService
-    {
-        public string HashPassword(string password)
-        {
-            return BCrypt.Net.BCrypt.HashPassword(password);
-        }
-        public bool VerifyPassword(string password, string hashedPassword)
-        {
-            return BCrypt.Net.BCrypt.Verify(password, hashedPassword);
-        }
-
-    }
-}
