+++ TTE.Commons/Services/SecurityService.cs
@@ -1,16 +1,13 @@
-﻿using System;
-using System.IdentityModel.Tokens.Jwt;
+﻿using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Text;
 using Microsoft.IdentityModel.Tokens;
-using TTE.Application.Interfaces;
-using TTE.Commons;
 
-namespace TTE.Application.Services
+namespace TTE.Commons.Services
 {
-    public class JwtService: IJwtService
+    public class SecurityService : ISecurityService
     {
-        public string GenerateToken(string username, string role)
+        public string GenerateToken(string username, string role, int id)
         {
             var key = Encoding.UTF8.GetBytes(EnvVariables.JWT_SECRET);
             var tokenHandler = new JwtSecurityTokenHandler();
@@ -18,6 +15,7 @@ namespace TTE.Application.Services
             var claims = new List<Claim>
             {
                 new Claim(ClaimTypes.Name, username),
+                new Claim("Uid", id.ToString()),
                 new Claim(ClaimTypes.Role, role),
                 new Claim(JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString())
             };
@@ -35,6 +33,17 @@ namespace TTE.Application.Services
 
             var token = tokenHandler.CreateToken(tokenDescriptor);
             return tokenHandler.WriteToken(token);
+        }
+
+
+
+        public string HashPassword(string password)
+        {
+            return BCrypt.Net.BCrypt.HashPassword(password);
+        }
+        public bool VerifyPassword(string password, string hashedPassword)
+        {
+            return BCrypt.Net.BCrypt.Verify(password, hashedPassword);
         }
