+++ TTE.Commons/SystemConstants.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TTE.Commons
+{
+    public static class SystemConstants
+    {
+        public const string MESSAGE_LOGIN_FAIL = "Invalid credentials";
+        public const string MESSAGE_LOGIN_SUCCESS = "Login successful";
+    }
+}
